Introduced `endpoint` as a base class of `client` and `server`.
Just moved client functionalites to the endpoint, so far.
